Updated install/upgradedb.php to use version.inc.php instead of VERSION.txt.
Also checks if file_get_contents (PHP 4.3+) and file_put_contents (PHP 5+) exist. If they do not exist, then the VERSION-DBCHECKED.txt is not read/written.

+++ trunk/install/upgradedb.php
@@ -15,18 +15,16 @@ function verifyUpgrade() {
 
 
 <?php
+@(include_once('../version.inc.php')) or die('version.inc.php was not found or could not be read. Make sure it exists in the VTCalendar folder and it defines a constant named "VERSION".');
+if (!defined("VERSION")) die('VERSION was not defined. Make sure version.inc.php defines a constant named "VERSION" (e.g. <code>define("VERSION", "2.3.0");</code>).');
 
-if (!file_exists("../VERSION.txt")) die("VERSION.txt was not found. Make sure it exists in the VTCalendar folder.");
-if (($version = file_get_contents("../VERSION.txt")) === false) die("VERSION.txt exists but could not be read. May not have read access to the file.");
-
-if (file_exists("../VERSION-DBCHECKED.txt")) {
+if (function_exists("file_get_contents") && file_exists("../VERSION-DBCHECKED.txt")) {
 	if (($dbVersionChecked = file_get_contents("../VERSION-DBCHECKED.txt")) === false) die("VERSION-DBCHECKED.txt exists but could not be. May not have read access to the file.");
-	if (trim($version) == trim($dbVersionChecked)) {
+	if (trim(VERSION) == trim($dbVersionChecked)) {
 		echo "<h1 style='color: red;'>Database Already Installed or Upgraded:</h1>"
 			."<p><a href='../VERSION-DBCHECKED.txt'><code>VERSION-DBCHECKED.txt</code></a> already matches <a href='../VERSION.txt'><code>VERSION.txt</code></a>.</p>"
 			."<p>If you would like to run this script, remove the <code>VERSION-DBCHECKED.txt</code> file in the VTCalendar folder.</p></body></html>";
 		exit;
-		//die("VERSION-DBCHECKED.txt already matches VERSION.txt.<br><br>If you would like to run this script, remove the VERSION-DBCHECKED.txt file in the VTCalendar folder.");
 	}
 }
 ?>
@@ -60,10 +58,12 @@ if (isset($Success)) {
 		?><div class="Success"><b>Success:</b> All upgrades were applied successfully!</div><?php
 	}
 	
-	$versionRecorded = (file_put_contents("../VERSION-DBCHECKED.txt", $version) !== false);
-	
-	if (!$versionRecorded) {
-		?><div class='Error'><b>Warning:</b> The <code>VERSION-DBCHECKED.txt</code> file could not be created/changed. To avoid people from accessing this page (and potentially compromising your database), copy the <code>VERSION.txt</code> file to <code>VERSION-DBCHECKED.txt</code>. On Linux the file is case-sensitive.</div><?php
+	if (function_exists("file_put_contents")) {
+		$versionRecorded = (file_put_contents("../VERSION-DBCHECKED.txt", VERSION) !== false);
+		
+		if (!$versionRecorded) {
+			?><div class='Error'><b>Warning:</b> The <code>VERSION-DBCHECKED.txt</code> file could not be created/changed. To avoid people from accessing this page (and potentially compromising your database), copy the <code>VERSION.txt</code> file to <code>VERSION-DBCHECKED.txt</code>. On Linux the file is case-sensitive.</div><?php
+		}
 	}
 }
 
@@ -108,8 +108,8 @@ elseif ($Submit_Preview && defined("DATABASE")) {
 					if ($changes == 0) {
 						echo '<script type="text/javascript">location.replace("upgradedb.php?Success=nochanges")</script>';
 					}
-					else {
-						$versionRecorded = (file_put_contents("../VERSION-DBCHECKED.txt", $version) !== false);
+					elseif (function_exists("file_put_contents")) {
+						$versionRecorded = (file_put_contents("../VERSION-DBCHECKED.txt", VERSION) !== false);
 						
 						if (!$versionRecorded) {
 							?><div class="Error"><b>Warning:</b> The <code>VERSION-DBCHECKED.txt</code> file could not be created/changed. To avoid people from accessing this page (and potentially compromising your database), copy the <code>VERSION.txt</code> file to <code>VERSION-DBCHECKED.txt</code>. On Linux the file is case-sensitive.</div><?php
